requirements.txt
service_identity==18.1.0
pydkim==0.3
pyopenssl==19.0.0
dnspython==1.16.0
twisted==19.2.1
requests_oauthlib=1.0.0
